requirements.txt
aniso8601==9.0.1
async-generator==1.10
attrs==22.2.0
certifi==2022.12.7
charset-normalizer==3.0.1
click==8.1.3
exceptiongroup==1.1.0
Flask==2.2.3
flask-restx==1.0.6
Flask-SQLAlchemy==3.0.3
greenlet==2.0.2
h11==0.14.0
idna==3.4
importlib-metadata==6.0.0
importlib-resources==5.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==4.17.3
MarkupSafe==2.1.2
outcome==1.2.0
packaging==23.0
Pillow==9.4.0
pkgutil_resolve_name==1.3.10
psycopg2-binary==2.9.5
pyrsistent==0.19.3
PySocks==1.7.1
python-dotenv==1.0.0
pytz==2022.7.1
requests==2.28.2
selenium==4.8.2
sniffio==1.3.0
sortedcontainers==2.4.0
SQLAlchemy==2.0.4
tqdm==4.64.1
trio==0.22.0
trio-websocket==0.9.2
typing_extensions==4.5.0
urllib3==1.26.14
webdriver-manager==3.8.5
Werkzeug==2.2.3
wsproto==1.2.0
zipp==3.15.0
